imported division to try and improve share/hr calc

+++ min-mon.py
@@ -279,8 +279,8 @@ def getEthminerData():
 
 	n=0
 	while n < i:
-		gpuTemps.append(gpu_temp_fanspeed[n*2])
-		gpuFanSpeeds.append(gpu_temp_fanspeed[(n*2)+1])
+		gpuTemps.append(int(gpu_temp_fanspeed[n*2]))
+		gpuFanSpeeds.append(int(gpu_temp_fanspeed[(n*2)+1]))
 		gpuDetails.append( gpuHashRates[n] + "," + gpuTemps[n] + "," + gpuFanSpeeds[n] )
 		n=n+1
